requirements.txt
Django==2.1.7
django-crispy-forms==1.9.1
Pillow==7.1.2
pytz==2020.1
